Update-Component
Update the original component with the new changes and delete the clone component riskAssessmentFlow2
