+++ force-app/main/default/lwc/riskAssessmentFlow/riskAssessmentFlow.css
@@ -5,4 +5,3 @@
 .spinner-hide{
     display: none;
 }
-
