--- force-app/main/default/lwc/riskAssessmentFlow2/riskAssessmentFlow2.css
@@ -1,7 +1,0 @@
-.spinner-show{
-    display: inline-block;
-}
-
-.spinner-hide{
-    display: none;
-}
