--- force-app/main/default/lwc/riskAssessmentFlow2/riskAssessmentFlow2.js-meta.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
-    <apiVersion>48.0</apiVersion>
-    <isExposed>true</isExposed>
-    <targets>
-        <target>lightning__AppPage</target>
-        <target>lightning__RecordPage</target>
-        <target>lightning__HomePage</target>
-        <target>lightningCommunity__Page</target>
-        <target>lightningCommunity__Default</target>
-    </targets>
-    <targetConfigs>
-        <targetConfig targets="lightning__HomePage">
-            <property name="cardTitle" label="Title" type="String"/>
-            <property name="riskAssessmentDefinitionName" label="Risk Assessment Def Name" type="String"/>
-        </targetConfig>
-        <targetConfig targets="lightning__AppPage, lightning__RecordPage">
-            <property name="cardTitle" label="Title" type="String"/>
-            <property name="riskAssessmentDefinitionName" label="Risk Assessment Def Name" type="String"/>
-            <supportedFormFactors>
-                <supportedFormFactor type="Large"/>
-                <supportedFormFactor type="Small"/>
-            </supportedFormFactors>
-        </targetConfig>
-        <targetConfig targets="lightningCommunity__Default">
-            <property name="cardTitle" label="Title" type="String"/>
-            <property name="riskAssessmentDefinitionName" label="Risk Assessment Def Name" type="String"/>
-            <property name="recordId" type="String" label="Record Id" description="Automatically bind the page's record id to the component variable" default="{!recordId}"/>
-        </targetConfig>
-    </targetConfigs>
-</LightningComponentBundle>
